% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_id_column.R
\name{add_id_column}
\alias{add_id_column}
\title{Add id column}
\usage{
add_id_column(data, sample)
}
\arguments{
\item{data}{}

\item{sample}{}
}
\value{

}
\description{
Add id column
}
\author{
Shir Dekel
}
