% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afex_plot.R
\name{afex_plot_alignment_8}
\alias{afex_plot_alignment_8}
\title{Plot afex aov object}
\usage{
afex_plot_alignment_8(afex_aov, dv_label)
}
\arguments{
\item{afex_aov}{\code{afex_aov} object}

\item{dv_label}{DV label}
}
\value{
A ggplot object
}
\description{
Plot afex aov object
}
\author{
Shir Dekel
}
