% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment8.R
\docType{package}
\name{alignment8}
\alias{alignment8}
\alias{alignment8-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
alignment8
A data package for alignment8.

Use \code{data(package='alignment8')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{plot_simulation}
\link{power}
}
