% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment8.R, R/alignment8data.R
\docType{package}
\name{alignment8data}
\alias{alignment8data}
\alias{alignment8data-package}
\title{Package Title}
\description{
A description of the data package

A description of the data package
}
\details{
alignment8data
A data package for alignment experiment 8.

Use \code{data(package='alignment8data')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.

alignment8data
A data package for alignment experiment 8.

Use \code{data(package='alignment8data')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{data}

\link{data}
}
