% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_other_alignment_8.R
\name{clean_data_other_alignment_8}
\alias{clean_data_other_alignment_8}
\title{Clean other data}
\usage{
clean_data_other_alignment_8(data_raw_prep, main_stage)
}
\value{

}
\description{
Clean other data
}
\author{
Shir Dekel
}
