% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_project_allocation.R
\name{clean_project_allocation}
\alias{clean_project_allocation}
\title{Convert responses from JSON and pivot}
\usage{
clean_project_allocation(data_raw_jspsych_columns_unselected)
}
\value{

}
\description{
Convert responses from JSON and pivot
}
\author{
Shir Dekel
}
