% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment8.R
\docType{data}
\name{data}
\alias{data}
\title{Descriptive data title}
\format{
a \code{rowwise_df} containing the following fields:
\describe{
\item{time_elapsed}{}
}
}
\source{
The data was collected using Prolific.
}
\description{
Detailed description of the data
}
\seealso{
\link{alignment8data}
}
