% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_business_name.R
\name{get_business_name}
\alias{get_business_name}
\title{Get business name
Names generated from namelix.com}
\usage{
get_business_name()
}
\value{

}
\description{
Get business name
Names generated from namelix.com
}
\author{
Shir Dekel
}
