% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_simulation_results.R
\name{get_data_simulation_results}
\alias{get_data_simulation_results}
\title{Get raw simulated data
For looping

Naming scheme: where the effect is, what effect is being compared}
\usage{
get_data_simulation_results(data, pb)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Get raw simulated data
For looping

Naming scheme: where the effect is, what effect is being compared
}
\author{
Shir Dekel
}
