% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_alignment_8.R
\name{get_df_alignment_8}
\alias{get_df_alignment_8}
\title{Get simulation data frame}
\usage{
get_df_alignment_8(n = 8, npv_amount)
}
\arguments{
\item{n}{Needs to be divisible by 8}
}
\value{

}
\description{
Get simulation data frame
}
\author{
Shir Dekel
}
