% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_name_materials_alignment_8.R
\name{get_file_name_materials_alignment_8}
\alias{get_file_name_materials_alignment_8}
\title{Get alignment 8 screenshot file names}
\usage{
get_file_name_materials_alignment_8()
}
\value{

}
\description{
Get alignment 8 screenshot file names
}
\author{
Shir Dekel
}
