% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instructions_alignment_8.R
\name{get_instructions_alignment_8}
\alias{get_instructions_alignment_8}
\title{Get instructions for alignment 8}
\usage{
get_instructions_alignment_8()
}
\value{

}
\description{
Get instructions for alignment 8
}
\author{
Shir Dekel
}
