% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interstitial.R
\name{get_interstitial}
\alias{get_interstitial}
\title{Get interstitial 1 for alignment 8}
\usage{
get_interstitial(interstitial_number)
}
\value{

}
\description{
Get interstitial 1 for alignment 8
}
\author{
Shir Dekel
}
