% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intrinsic_feature_multipliers.R
\name{get_intrinsic_feature_multipliers}
\alias{get_intrinsic_feature_multipliers}
\title{Get intrinsic feature multipliers}
\usage{
get_intrinsic_feature_multipliers(npv)
}
\value{

}
\description{
Get intrinsic feature multipliers
}
\author{
Shir Dekel
}
