% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_main.R
\name{get_main}
\alias{get_main}
\title{Get main experiment trials for alignment 8}
\usage{
get_main(gambles, randomize_order = TRUE)
}
\arguments{
\item{gambles}{}

\item{randomize_order}{}
}
\value{

}
\description{
Get main experiment trials for alignment 8
}
\author{
Shir Dekel
}
