% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_alignment_8.R
\name{get_model_alignment_8}
\alias{get_model_alignment_8}
\title{Alignment 8 model

Get model with regression coefficients of Experiment 2 for the explicit
condition, no effects for the implicit condition, and the intercept and
residual variance of Experiment 2.}
\usage{
get_model_alignment_8(n = 8, npv_amount)
}
\value{

}
\description{
Alignment 8 model

Get model with regression coefficients of Experiment 2 for the explicit
condition, no effects for the implicit condition, and the intercept and
residual variance of Experiment 2.
}
\author{
Shir Dekel
}
