% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_post_experiment.R
\name{get_post_experiment}
\alias{get_post_experiment}
\title{Get post experiment trials Alignment Experiment 8}
\usage{
get_post_experiment()
}
\value{

}
\description{
Get post experiment trials Alignment Experiment 8
}
\author{
Shir Dekel
}
