% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power_curve_alignment_8.R
\name{get_power_curve_alignment_8}
\alias{get_power_curve_alignment_8}
\title{Get power curve}
\usage{
get_power_curve_alignment_8(data_power, data_label)
}
\value{

}
\description{
Get power curve
}
\author{
Shir Dekel
}
