% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_allocation.R
\name{get_project_allocation}
\alias{get_project_allocation}
\title{Get project allocation}
\usage{
get_project_allocation(projects, randomize_order)
}
\value{

}
\description{
Get project allocation
}
\author{
Shir Dekel
}
