% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_detail_alignment_8.R
\name{get_project_detail_alignment_8}
\alias{get_project_detail_alignment_8}
\title{Get alignment 8 project details}
\usage{
get_project_detail_alignment_8()
}
\value{

}
\description{
Get alignment 8 project details
}
\author{
Shir Dekel
}
