% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_detail_description.R
\name{get_project_detail_feature}
\alias{get_project_detail_feature}
\title{Get project detail features}
\usage{
get_project_detail_feature()
}
\value{

}
\description{
Get project detail features
}
\author{
Shir Dekel
}
