% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_value_high.R
\name{get_project_value_high}
\alias{get_project_value_high}
\title{Get high project values}
\usage{
get_project_value_high()
}
\value{

}
\description{
Get high project values
}
\author{
Shir Dekel
}
