% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prolific_id_reject.R
\name{get_prolific_id_reject}
\alias{get_prolific_id_reject}
\title{Get prolific ID rejections}
\usage{
get_prolific_id_reject(data_check, prolific_filter, prolific_filter_label)
}
\value{

}
\description{
Get prolific ID rejections
}
\author{
Shir Dekel
}
