% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reliability_explicit.R
\name{get_reliability_explicit}
\alias{get_reliability_explicit}
\title{Get explicit reliability}
\usage{
get_reliability_explicit(npv)
}
\value{

}
\description{
Get explicit reliability
}
\author{
Shir Dekel
}
