% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reliability_implicit.R
\name{get_reliability_implicit}
\alias{get_reliability_implicit}
\title{Get implicit reliability}
\usage{
get_reliability_implicit(npv)
}
\value{

}
\description{
Get implicit reliability
}
\author{
Shir Dekel
}
