% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simulation_summary_alignment_8.R
\name{get_simulation_summary_alignment_8}
\alias{get_simulation_summary_alignment_8}
\title{Power analysis}
\usage{
get_simulation_summary_alignment_8(
  n = 32,
  model_alignment_8,
  nsim,
  pb,
  npv_amount
)
}
\value{

}
\description{
Power analysis
}
\author{
Shir Dekel
}
