% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_text.R
\name{get_survey_text}
\alias{get_survey_text}
\title{Get survey text}
\usage{
get_survey_text(
  label_text,
  name,
  minlength = NULL,
  maxlength = NULL,
  required = NULL,
  size = NULL,
  prefix = NULL,
  suffix = NULL
)
}
\value{

}
\description{
Get survey text
}
\author{
Shir Dekel
}
