% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_list.R
\name{multi_list}
\alias{multi_list}
\title{Generate HTML bullet point list}
\usage{
multi_list(x)
}
\arguments{
\item{x}{A vector of string to become become bullet points.}
}
\value{
A string.
}
\description{
Generate HTML bullet point list
}
\examples{
x <- c("a", "b", "c", "<br>")
multi_list(x)
}
