% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_business_name_latin.R
\name{mutate_business_name_latin}
\alias{mutate_business_name_latin}
\title{Mutate business name latin shuffle}
\usage{
mutate_business_name_latin(project_detail_alignment_8)
}
\value{

}
\description{
Mutate business name latin shuffle
}
\author{
Shir Dekel
}
