% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_project_table.R
\name{mutate_project_table}
\alias{mutate_project_table}
\title{Mutate project table}
\usage{
mutate_project_table(projects_business_name_latin)
}
\value{

}
\description{
Mutate project table
}
\author{
Shir Dekel
}
