% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_intrinsic_features.R
\name{pivot_intrinsic_features}
\alias{pivot_intrinsic_features}
\title{Pivot intrinsic features}
\usage{
pivot_intrinsic_features(data_raw_project_allocation_cleaned)
}
\value{

}
\description{
Pivot intrinsic features
}
\author{
Shir Dekel
}
