% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment8.R
\docType{data}
\name{plot_simulation}
\alias{plot_simulation}
\title{Descriptive data title}
\format{
a \code{gg} containing the following fields:
\describe{
\item{data}{}
\item{layers}{}
\item{scales}{}
\item{mapping}{}
\item{theme}{}
\item{coordinates}{}
\item{facet}{}
\item{plot_env}{}
\item{labels}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{alignment8}
\link{power}
}
