% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment8.R
\docType{data}
\name{power}
\alias{power}
\title{Descriptive data title}
\format{
a \code{list} containing the following fields:
\describe{
\item{power_curve}{}
\item{n}{}
\item{n_total}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{alignment8}
\link{plot_simulation}
}
