% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_interstitial.R
\name{rename_interstitial}
\alias{rename_interstitial}
\title{Rename interstitial

Fixes multiple interstitial_1 issue

If repeated, replace the second iteration of interstitial_1 with
interstitial_2}
\usage{
rename_interstitial(data_other)
}
\value{

}
\description{
Rename interstitial

Fixes multiple interstitial_1 issue

If repeated, replace the second iteration of interstitial_1 with
interstitial_2
}
\author{
Shir Dekel
}
